requirements.txt
langchain
deeplake
streamlit
python-dotenv
openai